requirements.txt
psycopg2-binary
flask
pandas
prometheus-client


